Doubleup

+++ src/main/java/parameters/NonScalable.java
@@ -2,5 +2,5 @@ package parameters;
 
 public interface NonScalable
 {
-    boolean matches(FixedParameter value, float epsilon);
+    boolean matches(FixedParameter value, double epsilon);
 }
