+++ src/main/java/parameters/Scalable.java
@@ -3,5 +3,5 @@ package parameters;
 public interface Scalable
 {
     float matches(FixedParameter value);
-    void expandBy(float value);
+    void expandBy(double value);
 }
